chore(webserver): sandbox assembly closure contracts, docs links, portable mutation script
- declare the sandbox same-origin assembly in the standalone gateway's
  component spec (requiredPorts), the process database pool contract
  (consumer + nested init/migrate/drift lifecycle relation with
  hostEvidence), closing the integration-closure gate the dependency
  refresh opened; link TECH-tunnel-frp-parity.md into the architecture
  document map (repository-docs gate)
- rewrite tools/mutation-check-webserver-materialization.sh portably
  (POSIX awk mutations with tmp+mv instead of GNU sed -i; sha256sum ->
  shasum -> openssl shim under the documented PORTABILITY:allow marker)
- PC/H5 test fixtures now carry the full required AppResponse shape
  (ownerType/tenantId/nginxConfigOverridden, ownerId in the operations
  column fixture) and the sandbox differential oracle tracks the
  sibling's current internal-api route authority with an explicit
  tripwire for an app-face reappearance
- refresh Cargo.lock to the latest compatible dependency versions and
  clear the remaining core/test warnings (unused variables/imports,
  trivial cast, certificate_fixture's dead path parameter)

+++ crates/sdkwork-webserver-core/src/config/source.rs
@@ -639,10 +639,7 @@ mod tests {
         assert_eq!(loaded.app.app_key, DEFAULT_APP_KEY);
         assert_eq!(loaded.app.virtual_hosts.len(), 1);
         let revision = loaded.revision.expect("single-file revision");
-        assert_eq!(
-            revision.size_bytes(),
-            fs::metadata(&toml).unwrap().len() as u64
-        );
+        assert_eq!(revision.size_bytes(), fs::metadata(&toml).unwrap().len());
     }
 
     #[test]
